:bug: :recycle: fix refactor: corrigindo badge em todo o lugar
