+++ frontend/src/components/Via/ViaCardSmall.vue
@@ -5,7 +5,7 @@
     </div>
     <div class="right-section">
       <div class="via-nome">{{ props.via.nome }}</div>
-      <GrauBadge :grauText="props.via.grau!" :extensaoText="props.via.extensao + 'm'" />
+      <GrauBadge :via="props.via"/>
     </div>
   </q-card-section>
 </template>
